refactor(pictograms): split into read + mutations modules (#191)
422-line `pictograms.ts` was the inconsistent outlier — `boards/` was
already split into `boards.read.ts` (query hooks + row narrowing) and
`boards.mutations.ts` (write hooks), with a 5-line `boards.ts` barrel
preserving consumer imports. Mirror that here:

- `pictograms.read.ts` (~80 lines): `rowToPictogram`, `pictogramsQueryKey`,
  `usePictograms`, `usePictogramsById`, `usePictogramsBySlug`.
- `pictograms.mutations.ts` (~315 lines): `pictogramToInsert`, the six
  mutation hooks, plus the private helpers (`patchPictogramInList`,
  `revokePictogramBlobs`) and `referencingBoardIds` they share.
- `pictograms.ts` becomes a 5-line barrel re-exporting both — all 13
  import sites keep working unchanged.
- Tests split the same way: `pictograms.read.test.ts` (rowToPictogram)
  and `pictograms.mutations.test.ts` (revokePictogramBlobs +
  pictogramToInsert round-trip). `pictograms.memo.test.tsx` was already
  read-side; left in place.

Pure code move — no behavior changes, no signature changes, no JSDoc
edits. Verified by full suite (339/339), typecheck, and lint.

Closes #191
